+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/data/factory/EnsoUuidFactory.scala
@@ -1,9 +1,9 @@
-package org.enso.languageserver.protocol.binary.factory
+package org.enso.languageserver.protocol.data.factory
 
 import java.util.UUID
 
 import com.google.flatbuffers.FlatBufferBuilder
-import org.enso.languageserver.protocol.binary.util.EnsoUUID
+import org.enso.languageserver.protocol.data.util.EnsoUUID
 
 object EnsoUuidFactory {
