+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/data/factory/ErrorFactory.scala
@@ -1,7 +1,7 @@
-package org.enso.languageserver.protocol.binary.factory
+package org.enso.languageserver.protocol.data.factory
 
 import com.google.flatbuffers.FlatBufferBuilder
-import org.enso.languageserver.protocol.binary.util.Error
+import org.enso.languageserver.protocol.data.util.Error
 
 object ErrorFactory {
