+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/data/factory/SessionInitResponseFactory.scala
@@ -1,7 +1,7 @@
-package org.enso.languageserver.protocol.binary.factory
+package org.enso.languageserver.protocol.data.factory
 
 import com.google.flatbuffers.FlatBufferBuilder
-import org.enso.languageserver.protocol.binary.session.SessionInitResponse
+import org.enso.languageserver.protocol.data.session.SessionInitResponse
 
 object SessionInitResponseFactory {
