+++ engine/language-server/src/main/scala/org/enso/languageserver/protocol/rpc/ClientController.scala
@@ -1,4 +1,4 @@
-package org.enso.languageserver.protocol
+package org.enso.languageserver.protocol.rpc
 
 import java.util.UUID
