+++ engine/language-server/src/main/scala/org/enso/languageserver/runtime/RuntimeFailureMapper.scala
@@ -2,7 +2,7 @@ package org.enso.languageserver.runtime
 
 import org.enso.jsonrpc._
 import org.enso.languageserver.filemanager.FileSystemFailureMapper
-import org.enso.languageserver.protocol.ErrorApi._
+import org.enso.languageserver.protocol.rpc.ErrorApi._
 import org.enso.languageserver.runtime.ExecutionApi._
 import org.enso.polyglot.runtime.Runtime.Api
