+++ engine/language-server/src/test/scala/org/enso/languageserver/websocket/data/OutboundMessageDecoder.scala
@@ -1,8 +1,8 @@
-package org.enso.languageserver.websocket.binary
+package org.enso.languageserver.websocket.data
 
 import java.nio.ByteBuffer
 
-import org.enso.languageserver.protocol.binary.envelope.OutboundMessage
+import org.enso.languageserver.protocol.data.envelope.OutboundMessage
 import org.enso.languageserver.util.binary.{BinaryDecoder, DecodingFailure}
 
 object OutboundMessageDecoder extends BinaryDecoder[OutboundMessage] {
