+++ engine/language-server/src/test/scala/org/enso/languageserver/websocket/data/VisualisationProtocolTest.scala
@@ -1,20 +1,20 @@
-package org.enso.languageserver.websocket.binary
+package org.enso.languageserver.websocket.data
 
 import java.nio.ByteBuffer
 import java.util.UUID
 
 import com.google.flatbuffers.FlatBufferBuilder
-import org.enso.languageserver.protocol.binary.envelope.{
+import org.enso.languageserver.protocol.data.envelope.{
   InboundPayload,
   OutboundMessage,
   OutboundPayload
 }
-import org.enso.languageserver.protocol.binary.executioncontext
+import org.enso.languageserver.protocol.data.executioncontext
 import org.enso.languageserver.runtime.VisualisationProtocol.{
   VisualisationContext,
   VisualisationUpdate
 }
-import org.enso.languageserver.websocket.binary.factory.{
+import org.enso.languageserver.websocket.data.factory.{
   InboundMessageFactory,
   SessionInitFactory
 }
