+++ engine/language-server/src/test/scala/org/enso/languageserver/websocket/rpc/ReceivesTreeUpdatesHandlerTest.scala
@@ -1,4 +1,4 @@
-package org.enso.languageserver.websocket
+package org.enso.languageserver.websocket.rpc
 
 import java.nio.file.{Files, Paths}
