+++ engine/language-server/src/test/scala/org/enso/languageserver/websocket/rpc/SessionManagementTest.scala
@@ -1,4 +1,5 @@
-package org.enso.languageserver.websocket
+package org.enso.languageserver.websocket.rpc
+
 import io.circe.literal._
 
 class SessionManagementTest extends BaseServerTest {
